migrate L.SheetGeometry and dependencies to typescript
(No functional changes)
